Isolando contexts em um estado global
