+++ pokedex/debug.log
@@ -1,0 +1,1 @@
+[0114/131435.987:ERROR:directory_reader_win.cc(43)] FindFirstFile: O sistema não pode encontrar o caminho especificado. (0x3)
