--- pokedex/src/Context/ContextPokedex.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const ContextPokedex = React.createContext()
-
-export default ContextPokedex
