--- pokedex/src/Context/ContextPokemonList.js
@@ -1,3 +1,0 @@
-import React from 'react'
-const ContextPokemonList = React.createContext()
-export default ContextPokemonList
