+++ pokedex/src/Global/GlobalStateContext.js
@@ -1,0 +1,5 @@
+import React from "react";
+
+const GlobalStateContext = React.createContext()
+
+export default GlobalStateContext;
